+++ R/Rzscore.R
@@ -6,11 +6,13 @@
 #' @param obj the FastWilcoxTest
 #' @title description of function Rzscore
 #' @export 
-if ( ! isGeneric('Rzscore') ){setGeneric('Rzscore', ## Name
+#if ( ! isGeneric('Rzscore') ){
+setGeneric('Rzscore', ## Name
 	function ( obj ) { 
 		standardGeneric('Rzscore')
 	}
-) }
+)
+# }
 
 setMethod('Rzscore', signature = c ('FastWilcoxTest'),
 	definition = function ( obj ) {
@@ -44,4 +46,26 @@ setMethod('Rzscore', signature = c ('FastWilcoxTest'),
 				)
 		), sparse =T)
 	invisible(obj)
+} )
+
+
+#' @name RzscoreAll
+#' @aliases RzscoreAll,FastWilcoxTest-method
+#' @rdname RzscoreAll-methods
+#' @docType methods
+#' @description calculate the z.score as the cpp function does it
+#' @param obj the FastWilcoxTest
+#' @title description of function Rzscore
+#' @export 
+#if ( ! isGeneric('RzscoreAll') ){
+setGeneric('RzscoreAll', ## Name
+	function ( obj ) { 
+		standardGeneric('RzscoreAll')
+	}
+) 
+#}
+
+setMethod('RzscoreAll', signature = c ('FastWilcoxTest'),
+	definition = function ( obj ) {
+		t(apply( obj@dat, 1 , function(d) { (d - mean(d)) / sd(d) } ))
 } )