+++ src/RcppExports.cpp
@@ -9,6 +9,11 @@
 
 using namespace Rcpp;
 
+#ifdef RCPP_USE_GLOBAL_ROSTREAM
+Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
+Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
+#endif
+
 // correlationCoefficient
 float correlationCoefficient(std::vector<double> X, std::vector<double> Y);
 static SEXP _FastWilcoxTest_correlationCoefficient_try(SEXP XSEXP, SEXP YSEXP) {
@@ -80,6 +85,42 @@ RcppExport SEXP _FastWilcoxTest_CorMatrixIDS(SEXP XSEXP, SEXP CMPSEXP, SEXP idsS
     UNPROTECT(1);
     return rcpp_result_gen;
 }
+// CorMatrixIDS_N
+NumericMatrix CorMatrixIDS_N(Eigen::MappedSparseMatrix<double> X, std::vector<double> CMP, std::vector<int> ids);
+static SEXP _FastWilcoxTest_CorMatrixIDS_N_try(SEXP XSEXP, SEXP CMPSEXP, SEXP idsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< Eigen::MappedSparseMatrix<double> >::type X(XSEXP);
+    Rcpp::traits::input_parameter< std::vector<double> >::type CMP(CMPSEXP);
+    Rcpp::traits::input_parameter< std::vector<int> >::type ids(idsSEXP);
+    rcpp_result_gen = Rcpp::wrap(CorMatrixIDS_N(X, CMP, ids));
+    return rcpp_result_gen;
+END_RCPP_RETURN_ERROR
+}
+RcppExport SEXP _FastWilcoxTest_CorMatrixIDS_N(SEXP XSEXP, SEXP CMPSEXP, SEXP idsSEXP) {
+    SEXP rcpp_result_gen;
+    {
+        Rcpp::RNGScope rcpp_rngScope_gen;
+        rcpp_result_gen = PROTECT(_FastWilcoxTest_CorMatrixIDS_N_try(XSEXP, CMPSEXP, idsSEXP));
+    }
+    Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
+    if (rcpp_isInterrupt_gen) {
+        UNPROTECT(1);
+        Rf_onintr();
+    }
+    bool rcpp_isLongjump_gen = Rcpp::internal::isLongjumpSentinel(rcpp_result_gen);
+    if (rcpp_isLongjump_gen) {
+        Rcpp::internal::resumeJump(rcpp_result_gen);
+    }
+    Rboolean rcpp_isError_gen = Rf_inherits(rcpp_result_gen, "try-error");
+    if (rcpp_isError_gen) {
+        SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
+        UNPROTECT(1);
+        Rf_error(CHAR(rcpp_msgSEXP_gen));
+    }
+    UNPROTECT(1);
+    return rcpp_result_gen;
+}
 // CorMatrix
 std::vector<double> CorMatrix(Eigen::SparseMatrix<double> X, std::vector<double> CMP);
 static SEXP _FastWilcoxTest_CorMatrix_try(SEXP XSEXP, SEXP CMPSEXP) {
@@ -96,6 +137,41 @@ RcppExport SEXP _FastWilcoxTest_CorMatrix(SEXP XSEXP, SEXP CMPSEXP) {
     {
         Rcpp::RNGScope rcpp_rngScope_gen;
         rcpp_result_gen = PROTECT(_FastWilcoxTest_CorMatrix_try(XSEXP, CMPSEXP));
+    }
+    Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
+    if (rcpp_isInterrupt_gen) {
+        UNPROTECT(1);
+        Rf_onintr();
+    }
+    bool rcpp_isLongjump_gen = Rcpp::internal::isLongjumpSentinel(rcpp_result_gen);
+    if (rcpp_isLongjump_gen) {
+        Rcpp::internal::resumeJump(rcpp_result_gen);
+    }
+    Rboolean rcpp_isError_gen = Rf_inherits(rcpp_result_gen, "try-error");
+    if (rcpp_isError_gen) {
+        SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
+        UNPROTECT(1);
+        Rf_error(CHAR(rcpp_msgSEXP_gen));
+    }
+    UNPROTECT(1);
+    return rcpp_result_gen;
+}
+// CorMatrix_N
+NumericMatrix CorMatrix_N(Eigen::SparseMatrix<double> X, std::vector<double> CMP);
+static SEXP _FastWilcoxTest_CorMatrix_N_try(SEXP XSEXP, SEXP CMPSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< Eigen::SparseMatrix<double> >::type X(XSEXP);
+    Rcpp::traits::input_parameter< std::vector<double> >::type CMP(CMPSEXP);
+    rcpp_result_gen = Rcpp::wrap(CorMatrix_N(X, CMP));
+    return rcpp_result_gen;
+END_RCPP_RETURN_ERROR
+}
+RcppExport SEXP _FastWilcoxTest_CorMatrix_N(SEXP XSEXP, SEXP CMPSEXP) {
+    SEXP rcpp_result_gen;
+    {
+        Rcpp::RNGScope rcpp_rngScope_gen;
+        rcpp_result_gen = PROTECT(_FastWilcoxTest_CorMatrix_N_try(XSEXP, CMPSEXP));
     }
     Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
     if (rcpp_isInterrupt_gen) {
@@ -298,6 +374,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// ShuffleMatrix
+Eigen::SparseMatrix<double> ShuffleMatrix(Eigen::SparseMatrix<double> X, int maxCols);
+RcppExport SEXP _FastWilcoxTest_ShuffleMatrix(SEXP XSEXP, SEXP maxColsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Eigen::SparseMatrix<double> >::type X(XSEXP);
+    Rcpp::traits::input_parameter< int >::type maxCols(maxColsSEXP);
+    rcpp_result_gen = Rcpp::wrap(ShuffleMatrix(X, maxCols));
+    return rcpp_result_gen;
+END_RCPP
+}
 // logFC
 double logFC(std::vector<double> A, std::vector<double> B);
 static SEXP _FastWilcoxTest_logFC_try(SEXP ASEXP, SEXP BSEXP) {
@@ -485,6 +573,18 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Eigen::SparseMatrix<double> >::type data(dataSEXP);
     Rcpp::traits::input_parameter< bool >::type display_progress(display_progressSEXP);
     rcpp_result_gen = Rcpp::wrap(ZScore(data, display_progress));
+    return rcpp_result_gen;
+END_RCPP
+}
+// ZScoreAll
+NumericMatrix ZScoreAll(Eigen::SparseMatrix<double> data, bool display_progress);
+RcppExport SEXP _FastWilcoxTest_ZScoreAll(SEXP dataSEXP, SEXP display_progressSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Eigen::SparseMatrix<double> >::type data(dataSEXP);
+    Rcpp::traits::input_parameter< bool >::type display_progress(display_progressSEXP);
+    rcpp_result_gen = Rcpp::wrap(ZScoreAll(data, display_progress));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -705,7 +805,80 @@ RcppExport SEXP _FastWilcoxTest_eDist3d(SEXP XSEXP, SEXP YSEXP, SEXP ZSEXP, SEXP
     UNPROTECT(1);
     return rcpp_result_gen;
 }
-
+// extract_proximity_oob
+NumericMatrix extract_proximity_oob(NumericMatrix pred, NumericMatrix prox, NumericMatrix inbag);
+RcppExport SEXP _FastWilcoxTest_extract_proximity_oob(SEXP predSEXP, SEXP proxSEXP, SEXP inbagSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type pred(predSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type prox(proxSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type inbag(inbagSEXP);
+    rcpp_result_gen = Rcpp::wrap(extract_proximity_oob(pred, prox, inbag));
+    return rcpp_result_gen;
+END_RCPP
+}
+// rollSum
+NumericMatrix rollSum(Eigen::SparseMatrix<double> X, int n);
+RcppExport SEXP _FastWilcoxTest_rollSum(SEXP XSEXP, SEXP nSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Eigen::SparseMatrix<double> >::type X(XSEXP);
+    Rcpp::traits::input_parameter< int >::type n(nSEXP);
+    rcpp_result_gen = Rcpp::wrap(rollSum(X, n));
+    return rcpp_result_gen;
+END_RCPP
+}
+// rollAreaSum
+NumericMatrix rollAreaSum(Eigen::SparseMatrix<double> X, std::vector<double> location, int funcID, int size);
+RcppExport SEXP _FastWilcoxTest_rollAreaSum(SEXP XSEXP, SEXP locationSEXP, SEXP funcIDSEXP, SEXP sizeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Eigen::SparseMatrix<double> >::type X(XSEXP);
+    Rcpp::traits::input_parameter< std::vector<double> >::type location(locationSEXP);
+    Rcpp::traits::input_parameter< int >::type funcID(funcIDSEXP);
+    Rcpp::traits::input_parameter< int >::type size(sizeSEXP);
+    rcpp_result_gen = Rcpp::wrap(rollAreaSum(X, location, funcID, size));
+    return rcpp_result_gen;
+END_RCPP
+}
+// sparse2SQLite_text_file
+void sparse2SQLite_text_file(Eigen::SparseMatrix<double> data, String file, char sep);
+static SEXP _FastWilcoxTest_sparse2SQLite_text_file_try(SEXP dataSEXP, SEXP fileSEXP, SEXP sepSEXP) {
+BEGIN_RCPP
+    Rcpp::traits::input_parameter< Eigen::SparseMatrix<double> >::type data(dataSEXP);
+    Rcpp::traits::input_parameter< String >::type file(fileSEXP);
+    Rcpp::traits::input_parameter< char >::type sep(sepSEXP);
+    sparse2SQLite_text_file(data, file, sep);
+    return R_NilValue;
+END_RCPP_RETURN_ERROR
+}
+RcppExport SEXP _FastWilcoxTest_sparse2SQLite_text_file(SEXP dataSEXP, SEXP fileSEXP, SEXP sepSEXP) {
+    SEXP rcpp_result_gen;
+    {
+        Rcpp::RNGScope rcpp_rngScope_gen;
+        rcpp_result_gen = PROTECT(_FastWilcoxTest_sparse2SQLite_text_file_try(dataSEXP, fileSEXP, sepSEXP));
+    }
+    Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
+    if (rcpp_isInterrupt_gen) {
+        UNPROTECT(1);
+        Rf_onintr();
+    }
+    bool rcpp_isLongjump_gen = Rcpp::internal::isLongjumpSentinel(rcpp_result_gen);
+    if (rcpp_isLongjump_gen) {
+        Rcpp::internal::resumeJump(rcpp_result_gen);
+    }
+    Rboolean rcpp_isError_gen = Rf_inherits(rcpp_result_gen, "try-error");
+    if (rcpp_isError_gen) {
+        SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
+        UNPROTECT(1);
+        Rf_error(CHAR(rcpp_msgSEXP_gen));
+    }
+    UNPROTECT(1);
+    return rcpp_result_gen;
+}
 // toColNums
 std::vector<double> toColNums(Eigen::SparseMatrix<double> data);
 static SEXP _FastWilcoxTest_toColNums_try(SEXP dataSEXP) {
@@ -781,9 +954,12 @@ static int _FastWilcoxTest_RcppExport_validate(const char* sig) {
     if (signatures.empty()) {
         signatures.insert("float(*correlationCoefficient)(std::vector<double>,std::vector<double>)");
         signatures.insert("std::vector<double>(*CorMatrixIDS)(Eigen::MappedSparseMatrix<double>,std::vector<double>,std::vector<int>)");
+        signatures.insert("NumericMatrix(*CorMatrixIDS_N)(Eigen::MappedSparseMatrix<double>,std::vector<double>,std::vector<int>)");
         signatures.insert("std::vector<double>(*CorMatrix)(Eigen::SparseMatrix<double>,std::vector<double>)");
+        signatures.insert("NumericMatrix(*CorMatrix_N)(Eigen::SparseMatrix<double>,std::vector<double>)");
         signatures.insert("std::vector<double>(*CorNormalMatrix)(NumericMatrix,std::vector<double>)");
-
+        signatures.insert("NumericMatrix(*rollSum)(Eigen::SparseMatrix<double>,int)");
+        signatures.insert("NumericMatrix(*rollSumStart)(Eigen::SparseMatrix<double>,double,std::vector<double>)");
         signatures.insert("NumericMatrix(*LinLang)(Eigen::SparseMatrix<double>,std::vector<int>,int,double,bool)");
         signatures.insert("double(*logFC)(std::vector<double>,std::vector<double>)");
         signatures.insert("std::vector<int>(*minusOne)(std::vector<int>)");
@@ -795,6 +971,7 @@ static int _FastWilcoxTest_RcppExport_validate(const char* sig) {
         signatures.insert("std::vector<double>(*euclidian_distances)(std::vector<double>,std::vector<double>,bool)");
         signatures.insert("std::vector<double>(*euclidian_distances3d)(std::vector<double>,std::vector<double>,std::vector<double>,bool)");
         signatures.insert("std::vector<double>(*eDist3d)(std::vector<double>,std::vector<double>,std::vector<double>,int)");
+        signatures.insert("void(*sparse2SQLite_text_file)(Eigen::SparseMatrix<double>,String,char)");
         signatures.insert("std::vector<double>(*toColNums)(Eigen::SparseMatrix<double>)");
         signatures.insert("std::vector<double>(*ColNotZero)(Eigen::SparseMatrix<double>)");
     }
@@ -805,9 +982,12 @@ static int _FastWilcoxTest_RcppExport_validate(const char* sig) {
 RcppExport SEXP _FastWilcoxTest_RcppExport_registerCCallable() { 
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_correlationCoefficient", (DL_FUNC)_FastWilcoxTest_correlationCoefficient_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_CorMatrixIDS", (DL_FUNC)_FastWilcoxTest_CorMatrixIDS_try);
+    R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_CorMatrixIDS_N", (DL_FUNC)_FastWilcoxTest_CorMatrixIDS_N_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_CorMatrix", (DL_FUNC)_FastWilcoxTest_CorMatrix_try);
+    R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_CorMatrix_N", (DL_FUNC)_FastWilcoxTest_CorMatrix_N_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_CorNormalMatrix", (DL_FUNC)_FastWilcoxTest_CorNormalMatrix_try);
-
+    R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_rollSum", (DL_FUNC)_FastWilcoxTest_rollSum_try);
+    R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_rollSumStart", (DL_FUNC)_FastWilcoxTest_rollSumStart_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_LinLang", (DL_FUNC)_FastWilcoxTest_LinLang_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_logFC", (DL_FUNC)_FastWilcoxTest_logFC_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_minusOne", (DL_FUNC)_FastWilcoxTest_minusOne_try);
@@ -819,6 +999,7 @@ RcppExport SEXP _FastWilcoxTest_RcppExport_registerCCallable() {
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_euclidian_distances", (DL_FUNC)_FastWilcoxTest_euclidian_distances_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_euclidian_distances3d", (DL_FUNC)_FastWilcoxTest_euclidian_distances3d_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_eDist3d", (DL_FUNC)_FastWilcoxTest_eDist3d_try);
+    R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_sparse2SQLite_text_file", (DL_FUNC)_FastWilcoxTest_sparse2SQLite_text_file_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_toColNums", (DL_FUNC)_FastWilcoxTest_toColNums_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_ColNotZero", (DL_FUNC)_FastWilcoxTest_ColNotZero_try);
     R_RegisterCCallable("FastWilcoxTest", "_FastWilcoxTest_RcppExport_validate", (DL_FUNC)_FastWilcoxTest_RcppExport_validate);
@@ -828,9 +1009,12 @@ RcppExport SEXP _FastWilcoxTest_RcppExport_registerCCallable() {
 static const R_CallMethodDef CallEntries[] = {
     {"_FastWilcoxTest_correlationCoefficient", (DL_FUNC) &_FastWilcoxTest_correlationCoefficient, 2},
     {"_FastWilcoxTest_CorMatrixIDS", (DL_FUNC) &_FastWilcoxTest_CorMatrixIDS, 3},
+    {"_FastWilcoxTest_CorMatrixIDS_N", (DL_FUNC) &_FastWilcoxTest_CorMatrixIDS_N, 3},
     {"_FastWilcoxTest_CorMatrix", (DL_FUNC) &_FastWilcoxTest_CorMatrix, 2},
+    {"_FastWilcoxTest_CorMatrix_N", (DL_FUNC) &_FastWilcoxTest_CorMatrix_N, 2},
     {"_FastWilcoxTest_CorNormalMatrix", (DL_FUNC) &_FastWilcoxTest_CorNormalMatrix, 2},
-
+    {"_FastWilcoxTest_rollSum", (DL_FUNC) &_FastWilcoxTest_rollSum, 2},
+    {"_FastWilcoxTest_rollSumStart", (DL_FUNC) &_FastWilcoxTest_rollSumStart, 3},
     {"_FastWilcoxTest_LinLang", (DL_FUNC) &_FastWilcoxTest_LinLang, 5},
     {"_FastWilcoxTest_LogNorm", (DL_FUNC) &_FastWilcoxTest_LogNorm, 3},
     {"_FastWilcoxTest_NormalizeCells", (DL_FUNC) &_FastWilcoxTest_NormalizeCells, 3},
@@ -842,6 +1026,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_FastWilcoxTest_cppWilcoxTest", (DL_FUNC) &_FastWilcoxTest_cppWilcoxTest, 3},
     {"_FastWilcoxTest_StatTest", (DL_FUNC) &_FastWilcoxTest_StatTest, 6},
     {"_FastWilcoxTest_ZScore", (DL_FUNC) &_FastWilcoxTest_ZScore, 2},
+    {"_FastWilcoxTest_ZScoreAll", (DL_FUNC) &_FastWilcoxTest_ZScoreAll, 2},
     {"_FastWilcoxTest_MEAN_STD", (DL_FUNC) &_FastWilcoxTest_MEAN_STD, 1},
     {"_FastWilcoxTest_SQRT", (DL_FUNC) &_FastWilcoxTest_SQRT, 1},
     {"_FastWilcoxTest_collapse", (DL_FUNC) &_FastWilcoxTest_collapse, 3},
@@ -850,7 +1035,10 @@ static const R_CallMethodDef CallEntries[] = {
     {"_FastWilcoxTest_euclidian_distances", (DL_FUNC) &_FastWilcoxTest_euclidian_distances, 3},
     {"_FastWilcoxTest_euclidian_distances3d", (DL_FUNC) &_FastWilcoxTest_euclidian_distances3d, 4},
     {"_FastWilcoxTest_eDist3d", (DL_FUNC) &_FastWilcoxTest_eDist3d, 4},
-
+    {"_FastWilcoxTest_extract_proximity_oob", (DL_FUNC) &_FastWilcoxTest_extract_proximity_oob, 3},
+    {"_FastWilcoxTest_rollSum", (DL_FUNC) &_FastWilcoxTest_rollSum, 2},
+    {"_FastWilcoxTest_rollAreaSum", (DL_FUNC) &_FastWilcoxTest_rollAreaSum, 4},
+    {"_FastWilcoxTest_sparse2SQLite_text_file", (DL_FUNC) &_FastWilcoxTest_sparse2SQLite_text_file, 3},
     {"_FastWilcoxTest_toColNums", (DL_FUNC) &_FastWilcoxTest_toColNums, 1},
     {"_FastWilcoxTest_ColNotZero", (DL_FUNC) &_FastWilcoxTest_ColNotZero, 1},
     {"_FastWilcoxTest_RcppExport_registerCCallable", (DL_FUNC) &_FastWilcoxTest_RcppExport_registerCCallable, 0},