+++ src/toColNums.cpp
@@ -1,6 +1,8 @@
 // [[Rcpp::depends(RcppEigen)]]
 #include <RcppEigen.h>
 #include <math.h>
+
+
 using namespace Rcpp;
 typedef Eigen::SparseVector<double> SpVec;
 typedef SpVec::InnerIterator InIterVec;
@@ -11,6 +13,14 @@ typedef SpVec::InnerIterator InIterVec;
 // [[Rcpp::interfaces(r, cpp)]]
 
 
+//' @title toColNums returns a vector with column IDs
+//' @aliases toColNums,FastWilcoxTest-method
+//' @rdname toColNums
+//' @description using c++ to get the column IDS mapping to the @x values.
+//' @param data a sparse matrix
+//' @return a vector with nGene information
+//' @export
+//' @return a vector of col ids in the order of the @x vector
 // [[Rcpp::export]]
 std::vector<double> toColNums(Eigen::SparseMatrix<double> data) {
   std::vector<double> tmp( data.nonZeros() );
@@ -24,6 +34,7 @@ std::vector<double> toColNums(Eigen::SparseMatrix<double> data) {
   }
   return tmp;
 }
+
 
 //' @title ColNotZero returns the amount of not zero values in each column
 //' @aliases ColNotZero,FastWilcoxTest-method